feat: 세션 말고 주스탠드 상태관리로 유저정보 불러오기로 다 바꿈
